* Fix warnings in pgm/transport. * Add missing export pgm_drop_superuser(). * Fix pgm_tsi_hash() to use re-entrant pgm_print_tsi_r(). * Fix return type of pgm_tsi_equal(). * Update all send() return types to gssize. * Update Reed Solomon module to use unsigned types. * Consolidate all sendto()'s to use one function, and to poll() retry upon   non-network related related. * Update transport source with return value notes. * Send functions now only fail on rate control, data is always pushed into   the transmit window. * Fix major typo in NAK pipe processing. * Remove all shadowed variables from transport. * Send stats only update when the sendto() succeeded. * Cleanup non-descript retvals. * Add checks on all sends for actual sent length.

+++ trunk/openpgm/pgm/include/pgm/rxwi.h
@@ -148,7 +148,7 @@ int pgm_rxw_shutdown (pgm_rxw_t*);
 
 int pgm_rxw_push_fragment (pgm_rxw_t*, gpointer, gsize, guint32, guint32, struct pgm_opt_fragment*, pgm_time_t);
 
-int pgm_rxw_readv (pgm_rxw_t*, pgm_msgv_t**, guint, struct iovec**, guint);
+gsize pgm_rxw_readv (pgm_rxw_t*, pgm_msgv_t**, guint, struct iovec**, guint);
 
 /* from state checking */
 int pgm_rxw_mark_lost (pgm_rxw_t*, guint32);
